refactor: Move tre_node to `parser.cst`
