+++ parser/cst/tree_node.py
@@ -3,10 +3,10 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING
 
-from .str_region import StrRegion
+from ..str_region import StrRegion
 
 if TYPE_CHECKING:
-    from .tokens import Token
+    from ..tokens import Token
 
 
 @dataclass
